backend/requirements.txt
# Backend requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.7.0
pydantic-settings==2.2.0
sqlalchemy==2.0.27
asyncpg==0.30.0
redis==5.1.0
python-multipart==0.0.6
PyJWT==2.8.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.0
Pillow==11.1.0
boto3==1.34.0
python-dotenv==1.0.0
